% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpath.R
\docType{data}
\name{.mpath_locale}
\alias{.mpath_locale}
\title{Locale to be used for m-Path data}
\format{
An object of class \code{locale} of length 7.
}
\usage{
.mpath_locale
}
\value{
Return a locale to be used in \code{\link[readr:read_delim]{readr::read_delim()}} or friends.
}
\description{
Hard coded locale to be used for m-Path data
}
\keyword{internal}
