% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpath_example.R
\name{mpath_example}
\alias{mpath_example}
\title{Get path to m-Path example data}
\usage{
mpath_example(file = NULL)
}
\arguments{
\item{file}{the name of the file to be accessed. If NULL, the function will return a list of all the example files.}
}
\value{
a character string with the path to the m-Path example data
}
\description{
This function provides an easy way to access the m-Path example files.
}
\examples{

# Example 1: access 'example_basic.csv' data

mpath_example('example_basic.csv') # returns the full path to the file 'example_basic.csv'

# Example 2: list all the example files

mpath_example() # returns the example files as a vector

}
