% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpath.R
\name{read_meta_data}
\alias{read_meta_data}
\title{Read m-Path meta data}
\usage{
read_meta_data(meta_data)
}
\arguments{
\item{meta_data}{A string with the path to the meta data file}
}
\value{
A \link[tibble]{tibble} with the contents of the meta data file.
}
\description{
Internal function to read the meta data file for an m-Path file.
}
\keyword{internal}
