% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rate.R
\name{response_rate}
\alias{response_rate}
\title{Calculate response rate}
\usage{
response_rate(
  data,
  valid_col,
  participant_col,
  time_col = NULL,
  period_start = NULL,
  period_end = NULL
)
}
\arguments{
\item{data}{data frame with data}

\item{valid_col}{name of the column that stores whether the beep was answered or not}

\item{participant_col}{name of the column that stores the participant id (or equivalent)}

\item{time_col}{optional: name of the column that stores the time of the beep}

\item{period_start}{optional: period start}

\item{period_end}{optional: period end}
}
\value{
a data frame with the response rate for each participant, and the number of beeps used to calculate the response rate
}
\description{
Calculate response rate
}
