% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpathr-package.R
\docType{package}
\name{mpathr-package}
\alias{mpathr}
\alias{mpathr-package}
\title{mpathr: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Handling data from the m-path platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-path-io/mpathr}
  \item Report bugs at \url{https://github.com/m-path-io/mpathr/issues}
}

}
\author{
\strong{Maintainer}: Koen Niemeijer \email{koen.niemeijer@kuleuven.be} (\href{https://orcid.org/0000-0002-0816-534X}{ORCID})

Authors:
\itemize{
  \item Merijn Mestdagh \email{merijn.mestdagh@kuleuven.be} (\href{https://orcid.org/0000-0001-5077-861X}{ORCID})
}

}
\keyword{internal}
