% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpathr-package.R
\docType{package}
\name{mpathr-package}
\alias{mpathr}
\alias{mpathr-package}
\title{mpathr: Easily Handling Data from the ‘m-Path’ Platform}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for importing and cleaning Experience Sampling Method ('ESM') data collected via the 'm-Path' platform. The goal is to provide with a few utility functions to be able to read and perform some common operations in 'ESM' data collected through the 'm-Path' platform (\url{https://m-path.io/landing/}). Functions include raw data handling, format standardization, and basic data checks, as well as to calculate the response rate in data from 'ESM' studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-path.io}
  \item \url{https://github.com/m-path-io/mpathr}
  \item Report bugs at \url{https://github.com/m-path-io/mpathr/issues}
}

}
\author{
\strong{Maintainer}: Merijn Mestdagh \email{merijn.mestdagh@m-path.io} (\href{https://orcid.org/0000-0001-5077-861X}{ORCID})

Authors:
\itemize{
  \item Lara Navarrete \email{larann901@gmail.com}
  \item Koen Niemeijer \email{koen.niemeijer@kuleuven.be} (\href{https://orcid.org/0000-0002-0816-534X}{ORCID})
}

Other contributors:
\itemize{
  \item m-Path Software [copyright holder]
}

}
\keyword{internal}
