% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpath.R
\name{read_mpath}
\alias{read_mpath}
\title{Read m-Path data}
\usage{
read_mpath(file, meta_data)
}
\arguments{
\item{file}{A string with the path to the m-Path file}

\item{meta_data}{A string with the path to the meta data file}
}
\value{
A \link[tibble]{tibble} with the contents of the m-Path file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function reads an m-Path file into a dataframe.
}
\details{
Note that this function has been tested only with the version v.1.1. of the meta_data.
So it is advised to use that version of the meta_data file
(you can change the meta data version under "export version" in "Export data").
}
\examples{

# We can use the function mpath_examples to get the path to the example data
basic_path <- mpath_example(file ='example_basic.csv')
meta_path <- mpath_example('example_meta.csv')

data <- read_mpath(file = basic_path,
                meta_data = meta_path)

}
