% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpath.R
\name{read_mpath}
\alias{read_mpath}
\title{Read an m-Path file into a tibble}
\usage{
read_mpath(file)
}
\arguments{
\item{file}{A string with the path to the m-Path file}
}
\value{
A \link[tibble]{tibble} with the contents of the m-Path file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function reads an m-Path file into a tibble.
}
\details{
Some more details...
}
\examples{
# TODO: create working examples
read_mpath("path/to/file")
}
