% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpath.R
\name{read_mpath}
\alias{read_mpath}
\title{Read m-Path data}
\usage{
read_mpath(file, meta_data)
}
\arguments{
\item{file}{A string with the path to the m-Path file}

\item{meta_data}{A string with the path to the meta data file}
}
\value{
A \link[tibble]{tibble} with the contents of the m-Path file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function reads an m-Path file into a dataframe.
}
\details{
Some more details...
}
\examples{

# the files for the example data and meta data are inside the package
# the following lines find the path to the files inside the package and save them
example_data_path <- system.file("extdata", "example_data.csv",
                            package = "mpathr", mustWork = TRUE)

example_meta_path <- system.file("extdata", "example_meta.csv",
                            package = "mpathr", mustWork = TRUE)

data <- read_mpath(file = example_data_path,
        meta_data = example_meta_path)

}
